pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "training-ai"
version = "0.1.0"
description = "Using OpenAI to do something with sports training data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic-ai==1.43.0",
  "pydantic-ai-slim[web]==1.43",
  "python-dotenv>=1.2.1",
  "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
]

[project.scripts]
training-ai = "cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = [
  "cli",
  "config",
  "signal_sender",
  "summary_agent",
  "summary_api",
  "training_agent",
  "web",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
